+++ CHANGELOG.md
@@ -1,1 +1,2 @@
-* Fixed HidingBar_Options.toc
+* Updated icon texture coord system for Masque.
+* Fixed scaling of own minimap button.